The viewer loader is now an explicit sub-zone. I replaced the flat loader module with a barrel at [packages/web/src/viewer/loader/index.ts](/Users/hal/Documents/VSCodeProjects/n-dx-internal/packages/web/src/viewer/loader/index.ts:1), moved the implementation to [packages/web/src/viewer/loader/data-loader.ts](/Users/hal/Documents/VSCodeProjects/n-dx-internal/packages/web/src/viewer/loader/data-loader.ts:1), and kept loader-only schema migration code inside the same subdirectory.
