+++ .rex/prd.json
@@ -3421,20 +3421,24 @@
         {
           "id": "90170684-ba0b-4cc6-b0a2-3dee79ebd88a",
           "title": "Fix structural in global (1 finding)",
-          "status": "pending",
+          "status": "completed",
           "level": "feature",
           "description": "- Three of five zones sit at or below cohesion 0.4; all three (web-3, polling, web-4) live under packages/web/src/viewer/ — the viewer package would benefit from barrel files at the polling/ and loader sub-directories to make public APIs explicit and reduce Louvain boundary ambiguity.",
           "priority": "high",
           "source": "sourcevision",
+          "startedAt": "2026-04-20T16:20:28.288Z",
+          "completedAt": "2026-04-20T16:20:28.288Z",
           "children": [
             {
               "id": "5831bfa9-0b2e-40b7-ac96-35bce716948f",
               "title": "Fix structural in global: Three of five zones sit at or below cohesion 0.4; all three (web-3, polling, web",
-              "status": "pending",
+              "status": "completed",
               "level": "task",
               "description": "- Three of five zones sit at or below cohesion 0.4; all three (web-3, polling, web-4) live under packages/web/src/viewer/ — the viewer package would benefit from barrel files at the polling/ and loader sub-directories to make public APIs explicit and reduce Louvain boundary ambiguity.",
               "priority": "high",
               "source": "sourcevision",
+              "startedAt": "2026-04-20T16:13:43.334Z",
+              "completedAt": "2026-04-20T16:20:28.231Z",
               "recommendationMeta": {
                 "findingHashes": [
                   "279c04baa6d5"
