+++ packages/web/src/viewer/loader/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Loader zone public interface.
+ *
+ * All cross-zone consumers should import from this barrel rather than
+ * individual implementation files.
+ */
+
+export {
+  getData,
+  onDataChange,
+  clearOnChange,
+  loadModules,
+  loadFromFiles,
+  detectMode,
+  startPolling,
+  stopPolling,
+} from "./data-loader.js";
